feature: react-map bevezetes II
