--- src/components/map/LayerStyle.json
@@ -1,27 +1,0 @@
-{
-    "startPozition" : {
-        "center" : [19.826587,47.9263058],
-        "zoom" : 12
-    },
-    "plan": {
-    "id": "planRoute-layer",
-    "type": "line", 
-    "paint": { "line-color": "#D4813A", "line-width": 6, "line-opacity": 0.7 }
-  },
-  "route": {
-    "id": "route-layer",
-    "type": "line", 
-    "paint": { 
-      "line-color": "#3A8D60", 
-      "line-opacity": 0.7, 
-      "line-width": [
-        "interpolate", ["linear"], ["zoom"],
-        8,  2,
-        14, 8
-    ]
-    }
-  },
-  "geojson": {
-    "type": "FeatureCollection", "features": [] 
-  }
-}
