--- src/components/map/LiveNavbar.css
@@ -1,75 +1,0 @@
-.navbar {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    padding: 0 16px;
-    background-color: var(--color-dark);
-}
-
-.menu-btn {
-    background: none;
-    border-color: var(--color-light);
-    border-radius: var(--radius-sm);
-    cursor: pointer;
-    padding: 4px;
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-}
-
-.menu-btn span {
-    display: block;
-    width: 24px;
-    height: 2px;
-    background: var(--color-light);
-    transition: background 0.2s;
-}
-
-.menu-btn:hover span {
-    background: var(--color-base);
-}
-
-.navbar-info {
-    flex: 1;
-    color: var(--color-text);
-    padding: 4px 10px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.navbar-status {
-    display: flex;
-    align-items: center;
-    gap: 16px;
-    flex-shrink: 0;
-}
-
-.status-led {
-    display: flex;
-    align-items: center;
-    gap: 3px;
-}
-
-.led {
-    display: block;
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    background: var(--color-dark);
-    border: 1px solid var(--color-light);
-}
-
-.led.active {
-    background: var(--color-moss);
-    box-shadow: 0 0 6px var(--color-moss);
-}
-
-.led.warning {
-    background: var(--color-warning);
-    box-shadow: 0 0 6px var(--color-warning);
-}
-
-.led.inactive {
-    background: var(--color-dark);
-}
