--- src/components/map/LiveStatus.css
@@ -1,7 +1,0 @@
-.statusDisplay { 
-    z-index: 100;
-    bottom: 10px;
-    left: 10px;
-    position: fixed; 
-    border-radius: var(--border-radius-md);
-};
